requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pyyaml>=6.0
transforms3d>=0.3.1
sympy>=1.9
opencv-python>=4.5.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Pinocchio和CasADi依赖（可选，用于高级IK求解）
#conda install pinocchio==3.1.0 -c conda-forge
casadi>=3.5.0
meshcat>=0.3.0

# Isaac Gym仿真依赖
# 注意：Isaac Gym需要从NVIDIA官方下载并手动安装
# 下载地址：https://developer.nvidia.com/isaac-gym
# Isaac Gym包含以下组件（会自动安装）：
# - isaacgym
# - isaacgymenvs  
torch>=1.10.0  # Isaac Gym需要PyTorch
tensorboard>=2.8.0  # 可选：用于训练日志记录 